refactor wrappers to be in their own directory, add more unit tests
The wrappers have been moved into their own directory so it's more clear
that they're for testing purposes and do not have to do with the actual
functionality of the components they used share files with.

More unit tests have been added for VMs/deployments, and existing unit
tests have been refactored to use the wrappers.