Fixes cluster-delete in Makefile (#253)
The command implied that the cluster-name was used to identify the
cluster to delete. This isn't the case (it needs to be the name
of the cluster as defined in the kubeconfig if set). As the kubeconfig
is generated per cluster, it's pretty safe to just point to it and not
specify a name.

Signed-off-by: John Harris <joharris@vmware.com>