cluster: Allow configurable APIServerPort
We were using port 443 as a hardcoded port in infrav1.APIEndpoint, which
was causing configuration issues, both for kubeadm and the resulting
kubeconfigs.

Here we add a APIServerPort method to the cluster scope, which allows us
to default to the correct API server port (6443) when a port is not
provided by the user.

The load balancers and NSGs are wired to use this method as well,
instead of hardcoded values.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>