feat: add failure domain support
Added the optional support for the failure domains to the provider as
as pecified by the CLuster API docs. Marked AvailabilityZone as
deprecated and added a warning  if its used.

Updated the `create-cluster` makefile target so that there is an option
not to deploy the manager to aid local debugging.

Relates To: #439