Add ability to autorun bootstrap/custom data script on Windows VMs
This commit adds to the WindowsConfiguration an AdditionalUnattendContent
configuration for automatically logging in as admin user and running the
CustomData/BootstrapData script on first boot.

This aligns the behavior when provisioning machines in the case of a
Windows VM on Azure with the behavior on AWS, where the CustomData
script is also autostarted on first boot. The same is the case for
Linux Machines on both Azure and AWS as well.

With this change, a PowerShell script can be placed on the machine and
will be autorun on first boot which allows for a very flexible way of
e.g. configuring the machine, installing additional software on the
machine, etc., all without the need for a custom Windows boot image.