refactor amcp and ammp to use transient error with requeue
This was done to take into account initial creation of
agent pools. Agent pools were making repeated 404 calls with
CreateOrUpdate when it was not needed. Now they wait for initialization
of the managed control plane prior to reconciling. This seems to be the
intent of initially reconciling the agent pools when the AKS instance is
first constructed.