api/v1beta1: introduce new ComputeGallery field
Azure has renamed Shared Image Gallery (SIG) to Azure Compute Gallery at
some point, so this commit reflects that by introducing new type which
also makes use of new ImagePlan struct to include plan information,
which makes it share more code with other types and also treats
ResourceGroup and SubscriptionID as string pointers as a preparation for
adding community galleries support, which do not require those fields to
be populated.

This is done as a separate step to better highlight what consists of an
actual community galleries support.

Refs #2267

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>