Makefile: put Calico manifest into artifacts directory and cache them
Right now running targets like 'generate' or 'generate-flavors' which
are often used when adding new flavors, downloads Calico release
manifests each time and they are 1GB in size, which is slow and
wasteful.

To speed things up and save bandwidth, this commit puts unpacked Calico
manifests in _artifacts directory as a build cache and uses Make file
target feature to only re-download the manifests if the files are
missing, which should only occur on fresh repository clone or when
Calico version changes, which then justifies re-downloading the archive.

To avoid downloading Calico archive multiple times when make targets are
executed in parallel (-j), we unpack entire manifests directory instead
of just desired manifests, which also makes the process simpler overall.

The manifests directory is right now <5MB, which is a lot (it's mostly
whitespace!) but I think acceptable.

Refs #2449

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>