Use azidentity instead of ADAL
ADAL is being deprecated, so use azidentity with an adapter so the new
authentication can work with v1 SDK clients.

Note that for certificate authentication, autorest uses the
`AZURE_CERTIFICATE_` prefix for environment variables whereas azidentity
uses `AZURE_CLIENT_CERTIFICATE_`. We make sure the latter is set to the
value of the former so as not to break upgrades.

Signed-off-by: Rafael Fonseca <r4f4rfs@gmail.com>