templates/flavors/flatcar: fix mounting etcd disk
Workaround for a CABPK bug with conversion of Ignition v2 to v3 as
described in https://github.com/kubernetes-sigs/cluster-api/issues/7679
has been added while Flatcar templates were still in development.

While it boots and does not produce any errors, it turns out that the
disk does not actually get mounted. This is because at the moment 'mounts'
CABPK field requires 'diskSetup.filesystems' to be populated for mapping
disk names to device path.

This commit fixes missing mount.

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>