Fix e2e config gob conversion to string
When converting the e2e config gob encoded bytes to string, one of those
bytes could be a `,` (comma), which leads to the next function accepting
the encoded e2e config to fail. This is due to the next function
accepting comma separated strings as arguments.

In order to keep the existing implementation as-is, i.e. pass comma
separated strings to the next function, we need to ensure that the
string representation of the gob encoded e2e config does not include
commas.

Using base64 encoding for the gob encoded e2e config resolves this
issue, as there is no comma in
[base64](https://datatracker.ietf.org/doc/html/rfc4648#section-4).

Signed-off-by: Michail Resvanis <mresvani@redhat.com>