Pin k8s test version to v1.26.3
Local run of CAPZ with Azure CNI v1 using K8s v1.26.3 is resulting in coredns pods not starting up.
Upon closer look at the v1.26.3 K8s cluster, it seems that the coredns pods are not able to resolve the DNS names.
This is a temporary pinning of K8s version to v1.26.3 to observe if CI picks up any issues.