feat(cluster): add failureDomain spec label
The goal of this commit is to allow capz users to specify
which failure domains are eligible for control plane rollouts.

There's a new label in AzureCluster.spec.failureDomain that can be
used to override the values of failureDomain.ControlPlane to false,
to prevent the control plane being deployed there.

The field is optional - if it's missing, all discovered failure domains
are announced in status as-is.

THERE IS NO BREAKING CHANGE TO CURRENT USERS.