Make subscriptionID mutable
Previously, the field was immutable. It was, and remains, an optional
field. The manager bootstrap credential was used when the field was
empty.

The credential is no longer supported, so the field is now effectively
required, because CAPZ fails to reconcile existing resources when the
field is empty.

Making the field mutable makes it possible to update existing resources
to the value previously stored in the manager bootstrap credential.