Add make verify/test to pre-push pre-commit config
This commit adds the same checks run in `make verify` without
duplicating any previous hook runs in pre-commit to the pre-commit
configuration. In addition, this commit adds `make test` to the pre-push
 stage of the pre-commit configuration. This will ensure `verify` and
`test` are successfully passing before the commit is pushed. This will
help reduce time and test resources by preventing `verify` and `test`
from failing on the pull request in GitHub.

Signed-off-by: Bryan Cox <brcox@redhat.com>